# md/pauth-l2-openapi-v1.yaml
openapi: 3.1.0
info:
  title: pauth API
  version: 1.0.0
  description: |
    OpenAPI specification for the implemented pauth authentication model.

    - L1: gateway forward-auth (`GET /api/verify`)
    - OAuth: app login (`/api/l2/authorize`, `/api/l2/token`, `/api/l2/userinfo`)
    - Admin: clients, users, L1 permissions, invites (`/api/admin/*`)

    All active pauth users may OAuth into registered clients. Per-app authorization is managed by each application.

    OAuth `redirect_uri`: any valid HTTPS URL (localhost HTTP for dev). No per-client URI registration.
servers:
  - url: https://auth.example.com
    description: Production
...
paths:
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List users with L1 flag
      security:
        - cookieSid: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, active, disabled, all] }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
...
components:
  securitySchemes:
    cookieSid:
      type: apiKey
      in: cookie
      name: sid
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    AdminUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string }
        status: { type: string }
        passkeyCount: { type: integer }
        hasPendingInvite: { type: boolean }
        l1Enabled: { type: boolean }
...